+++ Sanctuary.Census.Common/Objects/Collections/Experience.cs
@@ -11,6 +11,7 @@ namespace Sanctuary.Census.Common.Objects.Collections;
 /// <param name="ExperienceID">The ID of the experience point.</param>
 /// <param name="AwardTypeID">The ID of the award that this experience point grants.</param>
 /// <param name="Description">The description of this experience point.</param>
+/// <param name="LocalizedDescription">The description of this experience point.</param>
 /// <param name="XP">The amount of XP granted by this experience point.</param>
 /// <param name="IsNotableEvent">Indicates whether this experience point denotes an important player action.</param>
 [Collection]
@@ -18,7 +19,8 @@ public record Experience
 (
     [property:Key] uint ExperienceID,
     uint AwardTypeID,
-    LocaleString Description,
+    string? Description,
+    LocaleString? LocalizedDescription,
     decimal XP,
     bool IsNotableEvent
 ) : ISanctuaryCollection;